lab5/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
alembic==1.12.1
bcrypt==4.0.1
python-dotenv==1.0.0
pydantic[email]
websockets==12.0
celery==5.3.4
redis==5.0.1